+++ src/components/content/resource/helpers/history-entry.tsx
@@ -1,54 +1,64 @@
-import { ChevronRightIcon } from "@heroicons/react/outline";
+import { faFileLines } from "@fortawesome/free-regular-svg-icons";
+import { faChevronRight } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import cx from "classnames";
 import { ReactNode, useState } from "react";
-import { Details } from "./collapsible-data-list-details";
-import "./collapsible-data-list.scss";
-import { DocumentIcon } from "@/components/core/document-icon";
+import { DocumentButton } from "../../CCDA/document-button";
+import { useCCDAModal } from "../../CCDA/modal-ccda";
+import { DetailsCard } from "./details-card";
 
-export type CollapsibleDataListEntry = {
+export type DetailEntry = {
   label: string;
   value: ReactNode;
 };
 
-export type CollapsibleDataListProps = {
-  id: string;
-  date?: string;
-  title?: string;
-  subtitle?: string;
-  data: CollapsibleDataListEntry[];
-  hideEmpty?: boolean;
-  documentButton?: ReactNode;
+export type HistoryEntryProps = {
   binaryId?: string;
+  date?: string;
+  details: DetailEntry[];
+  hideEmpty?: boolean;
+  id: string;
+  subtitle?: string;
+  title?: string;
+  versionId?: string;
 };
 
-export const CollapsibleDataList = ({
+type Props = HistoryEntryProps & { resourceTypeTitle: string };
+
+export const HistoryEntry = ({
+  binaryId,
   date,
-  title,
-  subtitle,
-  data,
+  details,
   hideEmpty,
-  documentButton,
-}: CollapsibleDataListProps) => {
+  subtitle,
+  title,
+  resourceTypeTitle,
+}: Props) => {
   const [isDetailShown, setIsDetailShown] = useState(false);
+  const openCCDAModal = useCCDAModal();
 
   return (
-    <div
-      className="ctw-collapsible-data-list ctw-space-y-1"
-      data-zus-telemetry-namespace="CollapsibleDataList"
-    >
+    <div className="ctw-space-y-1" data-zus-telemetry-namespace="HistoryEntry">
       <DetailSummary
         date={date}
         title={title}
         subtitle={subtitle}
         isDetailShown={isDetailShown}
         setIsDetailShown={setIsDetailShown}
-        hasDocument={!!documentButton}
+        hasDocument={!!binaryId}
       />
       {isDetailShown && (
-        <Details
-          data={data}
+        <DetailsCard
+          details={details}
           hideEmpty={hideEmpty}
-          documentButton={documentButton}
+          documentButton={
+            binaryId && (
+              <DocumentButton
+                onClick={() => openCCDAModal(binaryId, resourceTypeTitle)}
+                text="Source Document"
+              />
+            )
+          }
         />
       )}
     </div>
@@ -76,7 +86,7 @@ const DetailSummary = ({
     onClick={() => setIsDetailShown(!isDetailShown)}
     data-zus-telemetry-namespace="DetailSummary"
     data-zus-telemetry-click={isDetailShown ? "Collapse" : "Expand"}
-    className="ctw-w-full ctw-cursor-pointer ctw-border-none ctw-bg-transparent ctw-p-0 ctw-text-base ctw-outline-none"
+    className="ctw-btn-clean"
   >
     <div className="ctw-flex ctw-items-center ctw-justify-between ctw-rounded-lg ctw-bg-bg-white ctw-p-3 ctw-text-left ctw-outline ctw-outline-1 ctw-outline-bg-dark">
       <div className="ctw-flex ctw-space-x-3">
@@ -90,14 +100,15 @@ const DetailSummary = ({
       </div>
       <div className="ctw-flex ctw-items-center ctw-space-x-3">
         {hasDocument && (
-          <DocumentIcon
-            className="ctw-fill-content-light hover:ctw-fill-content-light"
-            height={16}
+          <FontAwesomeIcon
+            icon={faFileLines}
+            className="ctw-h-5 ctw-text-content-light"
           />
         )}
         <div className="ctw-justify-right ctw-flex">
-          <ChevronRightIcon
-            className={cx("ctw-h-5 ctw-w-5 ctw-text-primary-dark", {
+          <FontAwesomeIcon
+            icon={faChevronRight}
+            className={cx("ctw-h-3 ctw-w-3 ctw-text-content-lighter", {
               "ctw-rotate-90": isDetailShown,
             })}
           />