+++ src/components/content/timeline/patient-timeline.tsx
@@ -1,12 +1,10 @@
 import cx from "classnames";
-import { useResourceDetailsDrawer } from "../resource/resource-details-drawer";
+import { usePatientEncounterDetailsDrawer } from "../timeline-2.0/helpers/modal-hooks";
 import { patientTimelineColumns } from "./patient-timeline-columns";
-import { CodingList } from "@/components/core/coding-list";
 import { useCTW } from "@/components/core/providers/ctw-provider";
 import { Table } from "@/components/core/table/table";
 import { usePatientEncounters } from "@/fhir/encounters";
-import { EncounterModel } from "@/fhir/models/encounter";
-import { capitalize, orderBy } from "@/utils/nodash";
+import { orderBy } from "@/utils/nodash";
 
 export type PatientTimelineProps = {
   className?: cx.Argument;
@@ -15,13 +13,7 @@ export type PatientTimelineProps = {
 export function PatientTimeline({ className }: PatientTimelineProps) {
   const patientEncounterQuery = usePatientEncounters();
   const { featureFlags } = useCTW();
-  const openDetails = useResourceDetailsDrawer({
-    header: (m) => `${m.periodStart} - ${m.periodEnd}`,
-    subHeader: (m) => m.typeDisplay,
-    getSourceDocument: true,
-    details: encounterData,
-  });
-
+  const openDetails = usePatientEncounterDetailsDrawer();
   const encounters = orderBy(
     patientEncounterQuery.data ?? [],
     [(encounter) => encounter.resource.period?.start ?? ""],
@@ -29,12 +21,7 @@ export function PatientTimeline({ className }: PatientTimelineProps) {
   );
 
   return (
-    <div
-      className={cx(
-        className,
-        "ctw-scrollable-pass-through-height ctw-overflow-hidden"
-      )}
-    >
+    <div className={cx(className, "ctw-scrollable-pass-through-height")}>
       <Table
         isLoading={patientEncounterQuery.isLoading}
         records={encounters}
@@ -44,21 +31,3 @@ export function PatientTimeline({ className }: PatientTimelineProps) {
     </div>
   );
 }
-
-const encounterData = (encounter: EncounterModel) => [
-  { label: "Period Start", value: encounter.periodStart },
-  { label: "Period End", value: encounter.periodEnd },
-  { label: "Status", value: capitalize(encounter.status) },
-  { label: "Class", value: encounter.class },
-  {
-    label: "Type",
-    value: encounter.typeCodings.length ? (
-      <CodingList codings={encounter.typeCodings} />
-    ) : undefined,
-  },
-  { label: "Location", value: encounter.location },
-  { label: "Participants", value: encounter.participants },
-  { label: "Reason", value: encounter.reason },
-  { label: "Diagnosis", value: encounter.diagnosis },
-  { label: "Discharge Disposition", value: encounter.dischargeDisposition },
-];