+++ src/components/core/list-box/list-box.tsx
@@ -13,19 +13,22 @@ import { FilterOptionSelect } from "@/components/core/filter-bar/filter-bar-type
 import { isFunction } from "@/utils/nodash";
 import "./list-box.scss";
 
-export type MinListBoxItem = {
+type ListBoxItem = {
   className?: cx.Argument;
+  divider?: false;
   display: string | ((status: ListBoxOptionStatus) => ReactNode);
   icon?: IconDefinition;
   key: string;
 };
 
-export type ListBoxProps<T> = {
+export type MinListBoxItem = ListBoxItem | { divider: true };
+
+export type ListBoxProps = {
   btnClassName?: cx.Argument;
   children?: ReactNode;
   defaultIndex?: number;
-  items: T[];
-  onChange: (index: number, item: T) => void;
+  items: MinListBoxItem[];
+  onChange: (index: number, item: ListBoxItem) => void;
   optionsClassName?: cx.Argument;
   useBasicStyles?: boolean;
 };
@@ -37,7 +40,7 @@ export type ListBoxOptionStatus = {
   selected?: boolean;
 };
 
-export function ListBox<T extends MinListBoxItem>({
+export function ListBox({
   useBasicStyles = false,
   btnClassName,
   children,
@@ -45,7 +48,7 @@ export function ListBox<T extends MinListBoxItem>({
   items,
   defaultIndex = 0,
   onChange,
-}: ListBoxProps<T>) {
+}: ListBoxProps) {
   const [selectedTabIndex, setSelectedTabIndex] = useState(defaultIndex);
   const selectedItem = items[selectedTabIndex] ?? {
     display: "Choose a selection",
@@ -56,7 +59,9 @@ export function ListBox<T extends MinListBoxItem>({
       <Listbox
         value={selectedTabIndex}
         onChange={(index: number) => {
-          onChange(index, items[index]);
+          const item = items[index];
+          if (item.divider) return;
+          onChange(index, item);
           setSelectedTabIndex(index);
         }}
       >
@@ -74,52 +79,64 @@ export function ListBox<T extends MinListBoxItem>({
           )}
         </Listbox.Button>
         <Listbox.Options className={cx(optionsClassName, "ctw-list-box")}>
-          {items.map((item, index) => (
-            <Listbox.Option key={item.key} value={index} as={Fragment}>
-              {({ active, selected }) => (
-                <li
-                  className={cx(
-                    "ctw-flex ctw-cursor-pointer ctw-justify-between ctw-px-3 ctw-py-2",
-                    "hover:ctw-bg-bg-lighter",
-                    item.className,
-                    {
-                      "ctw-text-medium": active || selected,
-                      "ctw-text-content-black": !(active || selected),
-                      "ctw-bg-bg-lighter": active,
-                    }
-                  )}
-                >
-                  <span
-                    className={cx("ctw-inline-flex ctw-align-middle", {
-                      "ctw-font-semibold": selected && !useBasicStyles,
-                    })}
+          {items.map((item, index) => {
+            if (item.divider) {
+              return <div key="divider" className="ctw-border-top" />;
+            }
+
+            return (
+              <Listbox.Option key={item.key} value={index} as={Fragment}>
+                {({ active, selected }) => (
+                  <li
+                    className={cx(
+                      "ctw-flex ctw-cursor-pointer ctw-justify-between ctw-px-3 ctw-py-2",
+                      "hover:ctw-bg-bg-lighter",
+                      item.className,
+                      {
+                        "ctw-text-medium": active || selected,
+                        "ctw-text-content-black": !(active || selected),
+                        "ctw-bg-bg-lighter": active,
+                      }
+                    )}
                   >
-                    {renderDisplay(item, { active, selected, listView: true })}
-                  </span>
-                  {!useBasicStyles && selected && !item.key.startsWith("_") && (
-                    <span className="ctw-inline-flex ctw-pb-0.5">
-                      <FontAwesomeIcon
-                        icon={faCheck}
-                        className="ctw-inline-block ctw-h-4 ctw-stroke-0 ctw-align-middle ctw-text-primary-dark"
-                      />
+                    <span
+                      className={cx("ctw-inline-flex ctw-align-middle", {
+                        "ctw-font-semibold": selected && !useBasicStyles,
+                      })}
+                    >
+                      {renderDisplay(item, {
+                        active,
+                        selected,
+                        listView: true,
+                      })}
                     </span>
-                  )}
-                </li>
-              )}
-            </Listbox.Option>
-          ))}
+                    {!useBasicStyles &&
+                      selected &&
+                      !item.key.startsWith("_") && (
+                        <span className="ctw-inline-flex ctw-pb-0.5">
+                          <FontAwesomeIcon
+                            icon={faCheck}
+                            className="ctw-inline-block ctw-h-4 ctw-stroke-0 ctw-align-middle ctw-text-primary-dark"
+                          />
+                        </span>
+                      )}
+                  </li>
+                )}
+              </Listbox.Option>
+            );
+          })}
         </Listbox.Options>
       </Listbox>
     </div>
   );
 }
 
-// TODO pull this out?
 // A helper function to render items for both the ListBox button and options.
-export function renderDisplay<T extends MinListBoxItem>(
+function renderDisplay<T extends MinListBoxItem>(
   item: T,
   status: ListBoxOptionStatus
 ) {
+  if (item.divider) return null;
   return (
     <MenuItem icon={item.icon}>
       {isFunction(item.display) ? item.display(status) : item.display}