+++ src/components/core/table/table-data-cell.tsx
@@ -5,13 +5,11 @@ import { MinRecordItem, TableColumn } from "./table-helpers";
 type TableColumnProps<T extends MinRecordItem> = {
   column: TableColumn<T>;
   record: T;
-  index: number;
 };
 
 export const TableDataCell = <T extends MinRecordItem>({
   column,
   record,
-  index,
 }: TableColumnProps<T>): JSX.Element => {
   const value = column.dataIndex
     ? (record[column.dataIndex] as unknown as ReactNode)
@@ -19,9 +17,8 @@ export const TableDataCell = <T extends MinRecordItem>({
 
   return (
     <td
-      key={column.title ?? index}
       className={cx(
-        { "ctw-font-medium ctw-text-content-black": index === 0 },
+        "ctw-text-content-black first:ctw-font-medium",
         column.className,
         // Added ctw-break-words because AbdominalReallyLongWordThatShouldBReakButDoeskadjlkajflkajsf doesn't wrap correclty without it.
         "ctw-hyphens-auto ctw-break-words"