+++ src/fhir/models/fhir-model.ts
@@ -1,10 +1,18 @@
 import { Basic, Resource } from "fhir/r4";
-import { SYSTEM_SUMMARY, SYSTEM_ZUS_PROFILE_ACTION } from "../system-urls";
+import {
+  SYSTEM_ENRICHMENT,
+  SYSTEM_SUMMARY,
+  SYSTEM_ZUS_PROFILE_ACTION,
+} from "../system-urls";
 import { ResourceMap } from "../types";
 import { find, startCase } from "@/utils/nodash";
 
 export abstract class FHIRModel<T extends fhir4.Resource> {
   public resource: T;
+
+  // This is a string that is used to identify the type of model and
+  // allows us to use a switch statement to determine the type of model.
+  abstract readonly kind: string;
 
   readonly includedResources?: ResourceMap;
 
@@ -22,6 +30,10 @@ export abstract class FHIRModel<T extends fhir4.Resource> {
 
   get id(): string {
     return this.resource.id || "";
+  }
+
+  get versionId(): string {
+    return this.resource.meta?.versionId || "";
   }
 
   get isArchived(): boolean {
@@ -50,6 +62,12 @@ export abstract class FHIRModel<T extends fhir4.Resource> {
         coding: [{ system: SYSTEM_ZUS_PROFILE_ACTION, code: profileAction }],
       },
     }) as Basic | undefined;
+  }
+
+  // Returns true if this resource is enriched with additional data
+  // from Zus enrichment.
+  isEnriched(): boolean {
+    return JSON.stringify(this.resource).includes(SYSTEM_ENRICHMENT);
   }
 
   // Returns a string that would setup this model.