+++ CHANGELOG.md
@@ -1,5 +1,44 @@
 # ctw-component-library
 
+## 1.1.1
+
+### Patch Changes
+
+- 4b55b511: Cleanup condition details/history drawer UI and fix a few minor bugs w.r.t "entered-in-error" conditions. This also sets up a generic details & history drawer to standardize other resource components.
+- eb02baaa: Add export for timeline.
+
+## 1.1.0
+
+### Minor Changes
+
+- bbe3c9ab: Add Multi-resource timeline component.
+
+## 1.0.0
+
+### Major Changes
+
+- 8f638ef9: Given where we are with maturity and stability we are now ready to release 1.0!
+
+## 0.51.1
+
+### Patch Changes
+
+- a2451acb: Tweak "Show Dismissed Records" filter to be below divider and to then show up as "Hide Dismissed Records" when it has been applied.
+
+## 0.51.0
+
+### Minor Changes
+
+- 0aeecb51: Refactor medication components. PatientMedications is now PatientMedicationsProfile which will show both the patient record and the outside medication tables.
+
+### Patch Changes
+
+- 0aeecb51: Patient medications now has a "view" selector where you can view "Current", "Past", or "All" medications.
+- 1b819762: Added ctw-scrollable-pass-through-height class to medications component.
+- 0aeecb51: Dismissed conditions will now render faded and still show their original status (e.g. "Active").
+- bd4e3503: Request records now appears in empty table state when patient has never made request to patient history api.
+- 0aeecb51: Medication tables will now show all information when stacked view for responsive/small width.
+
 ## 0.50.4
 
 ### Patch Changes